+++ Frends.Pdf.Create/Frends.Pdf.Create/Helpers/FileFontResolver.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
+using System.Threading;
 using Frends.Pdf.Create.Definitions;
 using PdfSharp.Fonts;
 
@@ -9,18 +10,23 @@ namespace Frends.Pdf.Create.Helpers;
 
 internal class FileFontResolver : IFontResolver
 {
-    private readonly List<FontMetadata> _fonts = [];
-    private readonly string[] _fontsLocations;
-    private readonly string _defaultFamilyName;
-    private readonly string _customFontsLocation;
+    private static AsyncLocal<List<FontMetadata>> _fonts;
+    private static AsyncLocal<string> _defaultFamilyName;
 
-    public FileFontResolver(string defaultFamilyName = "Arial", string customFontsLocation = null)
+    public FileFontResolver()
     {
-        _customFontsLocation = string.IsNullOrWhiteSpace(customFontsLocation) ? null : customFontsLocation;
-        _fontsLocations = GetFontsLocations();
-        _defaultFamilyName = string.IsNullOrWhiteSpace(defaultFamilyName) ? "Arial" : defaultFamilyName;
+        _fonts = new AsyncLocal<List<FontMetadata>>();
+        _defaultFamilyName = new AsyncLocal<string>();
+        _fonts.Value = [];
+    }
+
+    public static void Setup(string defaultFamilyName = "Arial", string customFontsLocation = null)
+    {
+        var fontsLocations = GetFontsLocations(customFontsLocation);
+        _defaultFamilyName.Value = string.IsNullOrWhiteSpace(defaultFamilyName) ? "Arial" : defaultFamilyName;
         List<string> fontsPaths = [];
-        foreach (var location in _fontsLocations)
+
+        foreach (var location in fontsLocations)
         {
             fontsPaths.AddRange(Directory.GetFiles(location, "*.ttf", SearchOption.AllDirectories));
         }
@@ -29,7 +35,11 @@ internal class FileFontResolver : IFontResolver
         {
             try
             {
-                _fonts.Add(new FontMetadata(file));
+                var newFont = new FontMetadata(file);
+                var alreadyExists = _fonts.Value.Any(f =>
+                    f.Name.Equals(newFont.Name, StringComparison.CurrentCultureIgnoreCase) &&
+                    f.IsBold == newFont.IsBold && f.IsItalic == newFont.IsItalic);
+                if (!alreadyExists) _fonts.Value.Add(newFont);
             }
             catch
             {
@@ -41,55 +51,46 @@ internal class FileFontResolver : IFontResolver
     public FontResolverInfo ResolveTypeface(string familyName, bool isBold, bool isItalic)
     {
         var font =
-            // try to get font we are looking for
-            _fonts.FirstOrDefault(f =>
+            // try to get the font we are looking for
+            _fonts.Value.FirstOrDefault(f =>
                 f.Name.Equals(familyName, StringComparison.CurrentCultureIgnoreCase)
                 && f.IsBold == isBold
                 && f.IsItalic == isItalic)
-            // try to get regular font from family
-            ?? _fonts.FirstOrDefault(f =>
+            // try to get a regular font from family
+            ?? _fonts.Value.FirstOrDefault(f =>
                 f.Name.Equals(familyName, StringComparison.CurrentCultureIgnoreCase)
                 && !f.IsBold
                 && !f.IsItalic)
             // try to get any font from family
-            ?? _fonts.FirstOrDefault(f => f.Name.Equals(familyName, StringComparison.CurrentCultureIgnoreCase))
-            // try to get regular fallback font
-            ?? _fonts.FirstOrDefault(f =>
-                f.Name.Equals(_defaultFamilyName, StringComparison.CurrentCultureIgnoreCase)
+            ?? _fonts.Value.FirstOrDefault(f => f.Name.Equals(familyName, StringComparison.CurrentCultureIgnoreCase))
+            // try to get a regular fallback font
+            ?? _fonts.Value.FirstOrDefault(f =>
+                f.Name.Equals(_defaultFamilyName.Value, StringComparison.CurrentCultureIgnoreCase)
                 && !f.IsBold
                 && !f.IsItalic)
-            // try to get any font from fallback family
-            ?? _fonts.FirstOrDefault(f =>
-                f.Name.Equals(_defaultFamilyName, StringComparison.CurrentCultureIgnoreCase))
+            // try to get any font from the fallback family
+            ?? _fonts.Value.FirstOrDefault(f =>
+                f.Name.Equals(_defaultFamilyName.Value, StringComparison.CurrentCultureIgnoreCase))
             ?? throw new Exception(
                 $"Font: {familyName} {(!isBold && !isItalic ? "regular," : string.Empty)}{(isBold ? "bold," : string.Empty)}{(isItalic ? "italic," : string.Empty)} couldn't be resolved");
-        return new FontResolverInfo(font.FileName);
+
+        return new FontResolverInfo(font.FullPath);
     }
 
-    public byte[] GetFont(string faceName)
+    public byte[] GetFont(string path)
     {
-        foreach (var location in _fontsLocations)
-        {
-            string fullPath = Path.Combine(location, faceName);
-            if (File.Exists(fullPath))
-            {
-                return File.ReadAllBytes(fullPath);
-            }
-        }
-
-        throw new Exception("Could not find font file");
+        return File.Exists(path) ? File.ReadAllBytes(path) : throw new Exception("Could not find font file");
     }
 
-    private string[] GetFontsLocations()
+    private static string[] GetFontsLocations(string customFontsLocation)
     {
+        var customLocation = string.IsNullOrWhiteSpace(customFontsLocation) ? null : customFontsLocation;
         List<string> result = [];
-        List<string> potentialPaths = [_customFontsLocation];
+        List<string> potentialPaths = [customLocation];
+
         if (OperatingSystem.IsWindows())
         {
-            potentialPaths.AddRange([
-                Environment.GetFolderPath(Environment.SpecialFolder.Fonts),
-                @"C:\Windows\fonts"
-            ]);
+            potentialPaths.Add(Environment.GetFolderPath(Environment.SpecialFolder.Fonts));
         }
         else if (OperatingSystem.IsLinux())
         {
@@ -98,7 +99,7 @@ internal class FileFontResolver : IFontResolver
                 "/usr/share/fonts",
                 "/usr/local/share/fonts",
                 Path.Combine(Environment.GetFolderPath(Environment.SpecialFolder.UserProfile), ".local/share/fonts"),
-                Path.Combine(Environment.GetFolderPath(Environment.SpecialFolder.UserProfile), ".fonts")
+                Path.Combine(Environment.GetFolderPath(Environment.SpecialFolder.UserProfile), ".fonts"),
             ]);
         }
         else throw new Exception("Unsupported operating system");